Combine common code
