+++ hinglishdistilbertfinetuning.py
@@ -8,13 +8,9 @@ import pandas as pd
 import tensorflow as tf
 import torch
 from IPython.display import clear_output
-
-
 from keras.preprocessing.sequence import pad_sequences
 from sklearn import preprocessing
 from sklearn.metrics import accuracy_score, f1_score, precision_recall_fscore_support
-
-
 from sklearn.model_selection import train_test_split
 from torch.utils.data import DataLoader, RandomSampler, SequentialSampler, TensorDataset
 from transformers import (
@@ -26,24 +22,14 @@ from transformers import (
     get_linear_schedule_with_warmup,
 )
 
-
-def all_the_important_configs(
-    batch_size=8,
-    attention_probs_dropout_prob=0.4,
-    learning_rate=5e-7,
-    adam_epsilon=1e-8,
-    hidden_dropout_prob=0.3,
-):
-    config = DistilBertConfig.from_json_file("output/config.json")
-    config.attention_probs_dropout_prob = attention_probs_dropout_prob
-    config.do_sample = True
-    config.num_beams = 500
-    config.hidden_dropout_prob = hidden_dropout_prob
-    config.repetition_penalty = 5
-    config.num_labels = 3
-
-    config
-    return batch_size, config, learning_rate, adam_epsilon
+from hinglishutils import (
+    all_the_important_configs,
+    check_for_gpu,
+    flat_accuracy,
+    flat_prf,
+    format_time,
+    load_sentences_and_labels,
+)
 
 
 def hinglishDistilBert(
@@ -72,6 +58,7 @@ def hinglishDistilBert(
         validation_labels,
     ) = load_masks_and_inputs(input_ids, labels, attention_masks)
     batch_size, config, learning_rate, adam_epsilon = all_the_important_configs(
+        "distilbert",
         batch_size,
         attention_probs_dropout_prob,
         learning_rate,
@@ -100,27 +87,6 @@ def hinglishDistilBert(
         num_warmup_steps=100,
         num_training_steps=total_steps,
     )
-
-    def flat_accuracy(preds, labels):
-        pred_flat = np.argmax(preds, axis=1).flatten()
-        labels_flat = labels.flatten()
-        return np.sum(pred_flat == labels_flat) / len(labels_flat)
-
-    def flat_prf(preds, labels):
-        pred_flat = np.argmax(preds, axis=1).flatten()
-        labels_flat = labels.flatten()
-        return precision_recall_fscore_support(
-            labels_flat, pred_flat, labels=[0, 1, 2], average="macro"
-        )
-
-    def format_time(elapsed):
-        """
-        Takes a time in seconds and returns a string hh:mm:ss
-        """
-
-        elapsed_rounded = int(round((elapsed)))
-
-        return str(datetime.timedelta(seconds=elapsed_rounded))
 
     def run_valid():
 
@@ -233,93 +199,6 @@ def save_model(full_output, model, tokenizer):
     tokenizer.save_pretrained(output_dir)
 
 
-def evaluate_final_text(tokenizer, MAX_LEN, model, device, le, final_name):
-    final_test_df = pd.read_json(final_name)
-    sentences = final_test_df["clean_text"]
-
-    input_ids = []
-
-    for sent in sentences:
-
-        encoded_sent = tokenizer.encode(
-            sent,
-            add_special_tokens=True,
-        )
-
-        input_ids.append(encoded_sent)
-
-    input_ids = pad_sequences(
-        input_ids, maxlen=MAX_LEN, dtype="long", truncating="post", padding="post"
-    )
-
-    attention_masks = []
-
-    for seq in input_ids:
-        seq_mask = [float(i > 0) for i in seq]
-        attention_masks.append(seq_mask)
-
-    prediction_inputs = torch.tensor(input_ids)
-    prediction_masks = torch.tensor(attention_masks)
-
-    batch_size = 32
-
-    prediction_data = TensorDataset(prediction_inputs, prediction_masks)
-    prediction_sampler = SequentialSampler(prediction_data)
-    prediction_dataloader = DataLoader(
-        prediction_data, sampler=prediction_sampler, batch_size=batch_size
-    )
-
-    open(f"{name}.log", "a").write(
-        "Predicting labels for {:,} valid sentences...\n".format(len(prediction_inputs))
-    )
-
-    model.eval()
-
-    predictions = []
-
-    for batch in prediction_dataloader:
-
-        batch = tuple(t.to(device) for t in batch)
-
-        b_input_ids, b_input_mask = batch
-
-        with torch.no_grad():
-
-            outputs = model(b_input_ids, attention_mask=b_input_mask)
-
-        logits = outputs[0]
-
-        logits = logits.detach().cpu().numpy()
-
-        predictions.append(logits)
-
-    open(f"{name}.log", "a").write("    DONE.\n")
-
-    flat_predictions = [item for sublist in predictions for item in sublist]
-    flat_predictions = np.argmax(flat_predictions, axis=1).flatten()
-    proba = [item for sublist in predictions for item in sublist]
-    preds = np.argmax(proba, axis=1).flatten()
-
-    output = le.inverse_transform(flat_predictions.tolist())
-    output_df = pd.DataFrame(
-        {
-            "Uid": list(final_test_df["uid"]),
-            "Sentiment": output,
-            "clean_text": list(final_test_df["clean_text"]),
-        }
-    )
-    output_df.to_csv(f"{name}-{final_name[:-5]}-output-df.csv")
-    proba = [item for sublist in predictions for item in sublist]
-    preds = np.argmax(proba, axis=1).flatten()
-    full_output = output_df
-    full_output["proba_negative"] = pd.DataFrame(proba)[0]
-    full_output["proba_neutral"] = pd.DataFrame(proba)[1]
-    full_output["proba_positive"] = pd.DataFrame(proba)[2]
-    full_output.to_csv(f"{name}-{final_name[:-5]}-full-output.csv")
-
-    return full_output
-
-
 def train_model(
     epochs,
     model,
@@ -402,113 +281,11 @@ def train_model(
     open(f"{name}.log", "a").write("Training complete!\n")
 
 
-def set_seed():
-    seed_val = 42
-
-    random.seed(seed_val)
-    np.random.seed(seed_val)
-    torch.manual_seed(seed_val)
-    torch.cuda.manual_seed_all(seed_val)
-
-
 def load_lm_model(config):
     model = DistilBertForSequenceClassification.from_pretrained("output", config=config)
     model.cuda()
     params = list(model.named_parameters())
     return model
-
-
-def make_dataloaders(
-    train_inputs,
-    train_masks,
-    train_labels,
-    batch_size,
-    validation_inputs,
-    validation_masks,
-    validation_labels,
-):
-
-    train_data = TensorDataset(train_inputs, train_masks, train_labels)
-    train_sampler = RandomSampler(train_data)
-    train_dataloader = DataLoader(
-        train_data, sampler=train_sampler, batch_size=batch_size
-    )
-
-    validation_data = TensorDataset(
-        validation_inputs, validation_masks, validation_labels
-    )
-    validation_sampler = SequentialSampler(validation_data)
-    validation_dataloader = DataLoader(
-        validation_data, sampler=validation_sampler, batch_size=batch_size
-    )
-    return train_dataloader, validation_dataloader
-
-
-def load_masks_and_inputs(input_ids, labels, attention_masks):
-
-    train_inputs, validation_inputs, train_labels, validation_labels = train_test_split(
-        input_ids, labels, random_state=2018, test_size=0.1
-    )
-
-    train_masks, validation_masks, _, _ = train_test_split(
-        attention_masks, labels, random_state=2018, test_size=0.1
-    )
-
-    train_inputs = torch.tensor(train_inputs)
-    validation_inputs = torch.tensor(validation_inputs)
-
-    train_labels = torch.tensor(train_labels)
-    validation_labels = torch.tensor(validation_labels)
-
-    train_masks = torch.tensor(train_masks)
-    validation_masks = torch.tensor(validation_masks)
-    return (
-        train_inputs,
-        train_masks,
-        train_labels,
-        validation_inputs,
-        validation_masks,
-        validation_labels,
-    )
-
-
-def create_attention_masks(input_ids):
-
-    attention_masks = []
-
-    for sent in input_ids:
-
-        att_mask = [int(token_id > 0) for token_id in sent]
-
-        attention_masks.append(att_mask)
-    return attention_masks
-
-
-def add_padding(tokenizer, input_ids):
-
-    MAX_LEN = 300
-
-    open(f"{name}.log", "a").write(
-        "\nPadding/truncating all sentences to %d values...\n" % MAX_LEN
-    )
-
-    open(f"{name}.log", "a").write(
-        '\nPadding token: "{:}", ID: {:}\n'.format(
-            tokenizer.pad_token, tokenizer.pad_token_id
-        )
-    )
-
-    input_ids = pad_sequences(
-        input_ids,
-        maxlen=MAX_LEN,
-        dtype="long",
-        value=0,
-        truncating="post",
-        padding="post",
-    )
-
-    open(f"{name}.log", "a").write("\nDone.")
-    return input_ids, MAX_LEN
 
 
 def tokenize_the_sentences(sentences):
@@ -530,24 +307,3 @@ def tokenize_the_sentences(sentences):
         input_ids.append(encoded_sent)
 
     return tokenizer, input_ids
-
-
-def check_for_gpu(device_name):
-
-    if device_name == "/device:GPU:0":
-        open(f"{name}.log", "a").write("Found GPU at: {}\n".format(device_name))
-        device = torch.device("cuda")
-    else:
-        raise SystemError("GPU device not found")
-    return device
-
-
-def load_sentences_and_labels():
-    train_df = pd.read_json("train.json")
-    test_df = pd.read_json("test.json")
-    sentences = train_df["clean_text"]
-    labels = train_df["sentiment"]
-    le = preprocessing.LabelEncoder()
-    le.fit(labels)
-    labels = le.transform(labels)
-    return sentences, labels, le
